\name{expext}
\alias{dexpext}
\alias{pexpext}
\alias{varexpext}
\alias{esexpext}
\title{Exponential extension distribution}
\description{Computes the pdf, cdf, value at risk and expected shortfall for the exponential extension distribution due to Nadarajah and Haghighi (2011) given by
\deqn{\begin{array}{ll}
&\displaystyle
f (x) =  a \lambda (1 + \lambda x)^{a - 1} \exp \left[ 1 - (1 + \lambda x)^a \right],
\\
&\displaystyle
F (x) =  1 - \exp \left[ 1 - (1 + \lambda x)^a \right],
\\
&\displaystyle
{\rm VaR}_p (X) = \frac {\left[ 1 - \log (1 - p) \right]^{1 / a} - 1}{\lambda},
\\
&\displaystyle
{\rm ES}_p (X) = -\frac {1}{\lambda} + \frac {1}{\lambda p} \int_0^p \left[ 1 - \log (1 - v) \right]^{1 / a} dv
\end{array}}
for \eqn{x > 0}, \eqn{0 < p < 1}, \eqn{a > 0}, the shape parameter and \eqn{\lambda > 0}, the scale parameter.}
\usage{
dexpext(x, lambda=1, a=1, log=FALSE)
pexpext(x, lambda=1, a=1, log.p=FALSE, lower.tail=TRUE)
varexpext(p, lambda=1, a=1, log.p=FALSE, lower.tail=TRUE)
esexpext(p, lambda=1, a=1)
}
\arguments{
  \item{x}{scaler or vector of values at which the pdf or cdf needs to be computed}
  \item{p}{scaler or vector of values at which the value at risk or expected shortfall needs to be computed}
  \item{lambda}{the value of the scale parameter, must be positive, the default is 1}
  \item{a}{the value of the shape parameter, must be positive, the default is 1}
  \item{log}{if TRUE then log(pdf) are returned}
  \item{log.p}{if TRUE then log(cdf) are returned and quantiles are computed for exp(p)}
  \item{lower.tail}{if FALSE then 1-cdf are returned and quantiles are computed for 1-p}
}
\value{An object of the same length as \code{x}, giving the pdf or cdf values computed at \code{x} or an object of the same length as \code{p}, giving the values at risk or expected shortfall computed at \code{p}.}
\references{Stephen Chan, Saralees Nadarajah & Emmanuel Afuecheta (2016). An R Package for Value at Risk and Expected Shortfall, Communications in Statistics - Simulation and Computation, 45:9, 3416-3434, \doi{10.1080/03610918.2014.944658}}
\author{Saralees Nadarajah}
\examples{x=runif(10,min=0,max=1)
dexpext(x)
pexpext(x)
varexpext(x)
esexpext(x)}
 
