\name{genbeta}
\alias{dgenbeta}
\alias{pgenbeta}
\alias{vargenbeta}
\alias{esgenbeta}
\title{Generalized beta distribution}
\description{Computes the pdf, cdf, value at risk and expected shortfall for the generalized beta distribution given by
\deqn{\begin{array}{ll}
&\displaystyle
f (x) = \frac {(x - c)^{a - 1} (d - x)^{b - 1}}{B (a, b) (d - c)^{a + b - 1}},
\\
&\displaystyle
F (x) = I_{\frac {x - c}{d - c}} (a, b),
\\
&\displaystyle
{\rm VaR}_p (X) = c + (d - c) I_p^{-1} (a, b),
\\
&\displaystyle
{\rm ES}_p (X) = c + \frac {d - c}{p} \int_0^p I_v^{-1} (a, b) dv
\end{array}}
for \eqn{c \leq x \leq d}, \eqn{0 < p < 1}, \eqn{a > 0}, the first shape parameter, \eqn{b > 0}, the second shape parameter, \eqn{-\infty < c < \infty}, the first location parameter, and \eqn{-\infty < c < d < \infty}, the second location parameter.}
\usage{
dgenbeta(x, a=1, b=1, c=0, d=1, log=FALSE)
pgenbeta(x, a=1, b=1, c=0, d=1, log.p=FALSE, lower.tail=TRUE)
vargenbeta(p, a=1, b=1, c=0, d=1, log.p=FALSE, lower.tail=TRUE)
esgenbeta(p, a=1, b=1, c=0, d=1)
}
\arguments{
  \item{x}{scaler or vector of values at which the pdf or cdf needs to be computed}
  \item{p}{scaler or vector of values at which the value at risk or expected shortfall needs to be computed}
  \item{c}{the value of the first location parameter, can take any real value, the default is zero}
  \item{d}{the value of the second location parameter, can take any real value but must be greater than c, the default is 1}
  \item{a}{the value of the first shape parameter, must be positive, the default is 1}
  \item{b}{the value of the second shape parameter, must be positive, the default is 1}
  \item{log}{if TRUE then log(pdf) are returned}
  \item{log.p}{if TRUE then log(cdf) are returned and quantiles are computed for exp(p)}
  \item{lower.tail}{if FALSE then 1-cdf are returned and quantiles are computed for 1-p}
}
\value{An object of the same length as \code{x}, giving the pdf or cdf values computed at \code{x} or an object of the same length as \code{p}, giving the values at risk or expected shortfall computed at \code{p}.}
\references{Stephen Chan, Saralees Nadarajah & Emmanuel Afuecheta (2016). An R Package for Value at Risk and Expected Shortfall, Communications in Statistics - Simulation and Computation, 45:9, 3416-3434, \doi{10.1080/03610918.2014.944658}}
\author{Saralees Nadarajah}
\examples{x=runif(10,min=0,max=1)
dgenbeta(x)
pgenbeta(x)
vargenbeta(x)
esgenbeta(x)}
 
