\encoding{utf8}

\name{ICSNP-package}
\alias{ICSNP-package}
\alias{ICSNP}
\docType{package}
\title{
\packageTitle{ICSNP}
}
\description{
\packageDescription{ICSNP}
}
\details{
\packageDESCRIPTION{ICSNP}

This package contains tools for nonparametric multivariate analysis, including the estimation of location and shape as well as some tests for location and independece.
Shape matrices from this package can be used as one of the scatter matrices needed in the package \code{\link[ICS:ICS-package]{ICS}} whereas the tests of this package
can be used for testing in the framework of invariant coordinates or independent components obtained from the package \code{\link[ICS:ICS-package]{ICS}}.
The parametric Hotelling's T test serves as a reference for the nonparametric location tests.

\packageIndices{ICSNP}
}

\author{
\packageAuthor{ICSNP}

Maintainer: \packageMaintainer{ICSNP}
}


\keyword{package}
\keyword{multivariate}
\keyword{nonparametric}
\seealso{
\code{\link[ICS:ICS-package]{ICS}}
}
