\name{logbeta}
\alias{dlogbeta}
\alias{plogbeta}
\alias{varlogbeta}
\alias{eslogbeta}
\title{Log beta distribution}
\description{Computes the pdf, cdf, value at risk and expected shortfall for the log beta distribution given by
\deqn{\begin{array}{ll}
&\displaystyle
f (x) = \frac {(\log d - \log c)^{1 - a - b}}{x B (a, b)}
(\log x - \log c)^{a - 1} (\log d - \log x)^{b - 1},
\\
&\displaystyle
F (x) = I_{\frac {\log x - \log c}{\log d - \log c}} (a, b),
\\
&\displaystyle
{\rm VaR}_p (X) = c \left( \frac {d}{c} \right)^{I_p^{-1} (a, b)},
\\
&\displaystyle
{\rm ES}_p (X) = \frac {c}{p} \int_0^p \left( \frac {d}{c} \right)^{I_v^{-1} (a, b)} dv
\end{array}}
for \eqn{0 < c \leq x \leq d}, \eqn{0 < p < 1}, \eqn{a > 0}, the first shape parameter, \eqn{b > 0}, the second shape parameter, \eqn{c > 0}, the first location parameter, and \eqn{d > 0}, the second location parameter.}
\usage{
dlogbeta(x, a=1, b=1, c=1, d=2, log=FALSE)
plogbeta(x, a=1, b=1, c=1, d=2, log.p=FALSE, lower.tail=TRUE)
varlogbeta(p, a=1, b=1, c=1, d=2, log.p=FALSE, lower.tail=TRUE)
eslogbeta(p, a=1, b=1, c=1, d=2)
}
\arguments{
  \item{x}{scaler or vector of values at which the pdf or cdf needs to be computed}
  \item{p}{scaler or vector of values at which the value at risk or expected shortfall needs to be computed}
  \item{c}{the value of the first location parameter, must be positive, the default is 1}
  \item{d}{the value of the second location parameter, must be positive and greater than c, the default is 2}
  \item{a}{the value of the first scale parameter, must be positive, the default is 1}
  \item{b}{the value of the second scale parameter, must be positive, the default is 1}
  \item{log}{if TRUE then log(pdf) are returned}
  \item{log.p}{if TRUE then log(cdf) are returned and quantiles are computed for exp(p)}
  \item{lower.tail}{if FALSE then 1-cdf are returned and quantiles are computed for 1-p}
}
\value{An object of the same length as \code{x}, giving the pdf or cdf values computed at \code{x} or an object of the same length as \code{p}, giving the values at risk or expected shortfall computed at \code{p}.}
\references{Stephen Chan, Saralees Nadarajah & Emmanuel Afuecheta (2016). An R Package for Value at Risk and Expected Shortfall, Communications in Statistics - Simulation and Computation, 45:9, 3416-3434, \doi{10.1080/03610918.2014.944658}}
\author{Saralees Nadarajah}
\examples{x=runif(10,min=0,max=1)
dlogbeta(x)
plogbeta(x)
varlogbeta(x)
eslogbeta(x)}
 
