% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/41_bondPricingUsingZspread.R
\name{pricingWithZspread}
\alias{pricingWithZspread}
\title{Calculates Bond Price using the given value of a Z-Spread and spot rates taken from the spots curve.}
\usage{
pricingWithZspread(cpns, spots, t, mv, n, zSprd)
}
\arguments{
\item{cpns}{A vector.}

\item{spots}{A vector.}

\item{t}{A vector.}

\item{mv}{A number.}

\item{n}{A number.}

\item{zSprd}{A number.}
}
\value{
Input values to six arguments \code{cpns},  \code{spots} ,\code{t},\code{mv}, \code{n} and \code{zSprd}.
}
\description{
Calculates Bond Price using the given value of a Z-Spread and spot rates taken from the spots curve.
}
\details{
According to information provided by Adams and Smith (2019), the method \code{pricingWithZspread()} is developed to compute Bond Price using the given value of a Z-Spread and spot rates taken from the spots curve. Here, \code{cpns} is vector of Coupon Payments,  \code{spots} is a vector of spot rates taken from the spots curve, \code{t} is a vector of number of years ranging from 1 to any specified number of years under consideration, \code{mv} is maturity value of the bond, \code{n} is number of years for which spots are available, and \code{zSprd} is given value of a Z-spread.
}
\examples{
pricingWithZspread(cpns=c(6,6),spots=c(0.0210,0.03635),t=c(1,2),mv=100,n=2,zSprd=0.023422)
pricingWithZspread(cpns=c(5,5,5),spots=c(0.0486,0.0495,0.0565),t=c(1,2,3),mv=100,n=3,zSprd=0.0234)
pricingWithZspread(cpns=c(rep(5,3)),spots=c(0.0486,0.0495,0.0565),t=c(1:3),mv=100,n=3,zSprd=0.0234)
pricingWithZspread(c(rep(5,3)),c(0.0486,0.0495,0.0565), c(1:3),100,3, zSprd=0.0234)
}
\references{
Adams,J.F. & Smith,D.J.(2019). Introduction to fixed-income valuation. In CFA Program Curriculum 2020 Level I Volumes 1-6. (Vol. 5, pp. 107-151). Wiley Professional Development (P&T). ISBN 9781119593577, \url{https://bookshelf.vitalsource.com/books/9781119593577}
}
\author{
MaheshP Kumar, \email{maheshparamjitkumar@gmail.com}
}
